refactor(cart): modularize cart components and extract logic into separate files
Extract cart components (CartItemList, CartSummary, EmptyCart, CheckoutForm, SuccessMessage) and types into separate files for better maintainability and reusability. Move form handling logic into a custom hook (useCheckoutForm) to centralize state management and improve code organization. This refactor enhances readability and reduces code duplication in the CartPage component.
