+++ src/components/cart/index.ts
@@ -1,0 +1,9 @@
+// Export des composants du panier
+export { default as CartItemList } from './CartItemList';
+export { default as CartSummary } from './CartSummary';
+export { default as EmptyCart } from './EmptyCart';
+export { default as CheckoutForm } from './CheckoutForm';
+export { default as SuccessMessage } from './SuccessMessage';
+
+// Export des types
+export * from './types';
